requirements.txt
Django==5.2.5
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
django-cors-headers==4.7.0
Pillow==11.3